+++ apps/app/src/storage/localstorage/Keys.ts
@@ -1,0 +1,3 @@
+export enum StorageKey {
+  Trips = "trips",
+}
