--- apps/app/src/utilities/global.ts
@@ -1,1 +1,0 @@
-export const leftPad = (n: number): string => n.toString().padStart(2, "0")
