+++ apps/app/src/utilities/index.ts
@@ -1,0 +1,1 @@
+export const leftPad = (n: number): string => n.toString().padStart(2, "0");
